backend/requirements.txt
asgiref==3.8.1
Django==4.2.20
django-cors-headers==4.7.0
djangorestframework==3.16.0
exceptiongroup==1.2.2
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
pytest==8.3.4
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.13.2
tzlocal==5.3
whitenoise==6.9.0
gunicorn>=21.2
